==> 1_data_preprocessing.ipynb <==
import pandas as pd
import numpy as np

crime = pd.read_csv("NYPD_Complaint_Data_Historic.csv")
crime

crime.info()

crime_renamed = crime.rename(columns = {
'CMPLNT_NUM': 'id',#
'CMPLNT_FR_DT': 'start_date_of_event',#
'CMPLNT_FR_TM': 'start_time_of_event',#
'CMPLNT_TO_DT': 'end_date_of_event',
'CMPLNT_TO_TM': 'end_time_of_event',
'ADDR_PCT_CD': 'precinct',
'RPT_DT': 'date_police_report',#
'KY_CD': 'offense_class',
'OFNS_DESC': 'description',#
'PD_CD': 'offense_key',
'PD_DESC': 'police_description',
'CRM_ATPT_CPTD_CD': 'crime_completion',
'LAW_CAT_CD': 'offense_level',
'BORO_NM': 'boro',#
'LOC_OF_OCCUR_DESC': 'spatial_loc',
'PREM_TYP_DESC': 'location_type',
'JURIS_DESC': 'juris_description',
'JURISDICTION_CODE': 'juris_code',
'PARKS_NM': 'greenspace_name',
'HADEVELOPT': 'NYCHA_name',
'HOUSING_PSA': 'dvlpmnt_lvl_code',
'X_COORD_CD': 'X_coord',
'Y_COORD_CD': 'y_coord',
'SUSP_AGE_GROUP': 'suspect_age',
'SUSP_RACE': 'suspect_race',
'SUSP_SEX': 'suspect_sex',
'TRANSIT_DISTRICT': 'trans_dstrct',
'Latitude': 'lat',#
'Longitude': 'lon',#
'Lat_Lon': 'lat_lon',
'PATROL_BORO': 'patrol_boro',
'STATION_NAME': 'trans_station_name',
'VIC_AGE_GROUP': 'victim_age',
'VIC_RACE': 'victim_race',
'VIC_SEX': 'victim_sex'
}).copy()

crime_renamed.to_csv("NYPD_Complaint_Renamed.csv", index = False)

crime_reduced = crime_renamed[[
    'id',
    'start_date_of_event',
    'start_time_of_event',
    'precinct',
    'date_police_report',
    'description',
    'police_description',
    'offense_level',
    'boro',
    'location_type',
    'juris_description',
    'lat',
    'lon',
    'trans_station_name',
    'victim_age',
    'victim_race',
    'victim_sex'
]]

crime_reduced.info()

crime_reduced.head()

crime_reduced.to_csv("NYPD_Complaint.csv", index = False)
