==> 2_data_cleaning.ipynb <==
import warnings
warnings.filterwarnings("ignore")

import pandas as pd
import numpy as np

crime = pd.read_csv("NYPD_Complaint.csv")

crime.info(verbose = True, show_counts = True)

crime_serious = crime.loc[(crime["offense_level"] == "FELONY") | (crime["offense_level"] == "MISDEMEANOR")].copy()
crime_serious

seventeen_crimes = crime_serious["description"].value_counts().head(17).index.to_list()
seventeen_crimes

crime_top17 = crime_serious.loc[crime_serious["description"].isin(seventeen_crimes)]
crime_top17

crime_top17.info(verbose = True, show_counts = True)

crime_top17.dropna(subset = ["start_date_of_event", "start_time_of_event", "lat", "lon"], inplace = True)

crime_top17 = crime_top17[(crime_top17["lat"] >= 40.217052) & (crime_top17["lat"] <= 41.053429)]

crime_top17 = crime_top17[(crime_top17["lon"] >= -74.742935) & (crime_top17["lon"] <= -73.538734)]

crime_top17[["month", "day", "year"]] = crime_top17["start_date_of_event"].str.split("/", expand = True)

crime_top17["year"] = crime_top17["year"].astype("int32")

crime_top17 = crime_top17[(crime_top17["year"] >= 2006) & (crime_top17["year"] < 2022)]

crime_top17 = crime_top17.apply(lambda x: x.str.title() if(x.dtype == "object") else x)

crime_top17["description"].value_counts()

crime_top17.replace(to_replace = "Assault 3 & Related Offenses", value = "Reckless Assault", regex = False, inplace = True)

crime_top17.replace(to_replace = "Criminal Mischief & Related Of", value = "Criminal Mischief", regex = False, inplace = True)

crime_top17.replace(to_replace = "Off. Agnst Pub Ord Sensblty &", value = "Offenses Against Public Order", regex = False, inplace = True)

crime_top17.replace(to_replace = "Offenses Against Public Admini", value = "Offenses Against Public Admin", regex = False, inplace = True)

crime_top17.replace(to_replace = "Asian / Pacific Islander", value = "Asian/Pacific Islander", regex = False, inplace = True)

crime_top17.head()

crime_top17.to_csv("NYPD_Complaint_Top.csv", index = False)

crime_streamlit = crime_top17.drop(columns = ["precinct", "date_police_report", "police_description", "offense_level", 
                            "location_type", "juris_description", "trans_station_name"])

crime_streamlit = crime_streamlit[crime_streamlit["victim_age"].isin(["<18", "18-24", "25-44", "45-64", "65+"])]

crime_streamlit = crime_streamlit[crime_streamlit["victim_race"].isin(["White Hispanic", "Black", "Black Hispanic", "White", 
                                                     "Asian/Pacific Islander", "American Indian/Alaskan Native"])]

crime_streamlit = crime_streamlit[crime_streamlit["victim_sex"].isin(["M", "F"])]

crime_streamlit

crime_streamlit.to_csv("NYPD_Complaint_Streamlit.csv", index = False)

crime_streamlit["lat_rnd"] = round(crime["lat"], 3)

crime_streamlit["lon_rnd"] = round(crime["lon"], 3)

crime_streamlit

crime_streamlit.to_csv("NYPD_Complaint_Folium.csv", index = False)
