==> 5_zipcode_lookup.ipynb <==
import pandas as pd
import numpy as np
import geopy

crime = pd.read_csv("Data/NYPD_Complaint_Folium.csv")
crime

crime_2021 = crime[crime["year"] == 2021]

crime_2021_10000 = crime_2021.sample(10000)

# def get_zipcode(df, geolocator, lat_field, lon_field):
#     location = geolocator.reverse((df[lat_field], df[lon_field]))
#     return location.raw["address"]["postcode"]

geolocator = geopy.Nominatim(user_agent = "geo2zip")

# crime_2021_10000["zipcodes"] = crime_2021_10000.apply(get_zipcode, axis = 1, geolocator = geolocator, 
#                                                       lat_field = "lat", lon_field = "lon")

def get_zipcode(geolocator, lat_field, lon_field):
    location = geolocator.reverse((lat_field, lon_field))
    try:
        result = location.raw["address"]["postcode"]
    except:
        result = "00000"
    return result

crime_2021_10000["zipcodes"] = crime_2021_10000.apply(lambda x: get_zipcode(geolocator, x.lat, x.lon), axis = 1)

crime_2021_10k = crime_2021_10000

crime_2020 = crime[crime["year"] == 2020]
crime_2020_10k = crime_2020.sample(10000)

crime_2020_10k["zipcodes"] = crime_2020_10k.apply(lambda x: get_zipcode(geolocator, x.lat, x.lon), axis = 1)

crime_2020_10k

crime_2019 = crime[crime["year"] == 2019]
crime_2019_10k = crime_2019.sample(10000)

crime_2019_10k["zipcodes"] = crime_2019_10k.apply(lambda x: get_zipcode(geolocator, x.lat, x.lon), axis = 1)

crime_2019_10k

crime_2018 = crime[crime["year"] == 2018]
crime_2018_10k = crime_2018.sample(10000)

crime_2018_10k["zipcodes"] = crime_2018_10k.apply(lambda x: get_zipcode(geolocator, x.lat, x.lon), axis = 1)

crime_2018_10k

crime_2017 = crime[crime["year"] == 2017]
crime_2017_10k = crime_2017.sample(10000)

crime_2017_10k["zipcodes"] = crime_2017_10k.apply(lambda x: get_zipcode(geolocator, x.lat, x.lon), axis = 1)

crime_2017_10k

crime_2016 = crime[crime["year"] == 2016]
crime_2016_10k = crime_2016.sample(10000)

crime_2016_10k["zipcodes"] = crime_2016_10k.apply(lambda x: get_zipcode(geolocator, x.lat, x.lon), axis = 1)
