==> 6_folium_map.ipynb <==
import pandas as pd
import numpy as np
import folium
from folium.plugins import HeatMap, MarkerCluster, HeatMapWithTime
import matplotlib.pyplot as plt

crime = pd.read_csv("Data/NYPD_Complaint_Folium.csv")
crime

crime.info()

crime_2021mar = crime[(crime["year"] == 2021) & (crime["month"] == 3)]

locations = list(zip(crime_2021mar["lat_rnd"], crime_2021mar["lon_rnd"]))

base_map = folium.Map(location = [40.730610, -73.935242], tiles = "cartodbpositron", zoom_start = 11)
# heatmap = HeatMap(locations, radius = 5, blur = 2).add_to(base_map)

cluster = MarkerCluster().add_to(base_map)

for x in crime.iterrows():
    folium.Marker([x[1].lat_rnd, x[1].lon_rnd]).add_to(cluster)

base_map

index_list = []
for i in range(2006, 2022):
    for j in range(1, 13):
        date = str(j) + "/" + str(i)
        index_list.append(date)

crime["month_year"] = crime["month"].astype(str) + "/" + crime["year"].astype(str)

crime["count"] = 1

weight_list = []

for x in crime["month_year"].sort_values().unique():
    weight_list.append(crime.loc[crime["month_year"] == x, ["lat_rnd", "lon_rnd", "count"]].groupby(["lat_rnd", "lon_rnd"]).sum().reset_index().values.tolist())

HeatMapWithTime(weight_list, radius = 10, index = index_list, auto_play = True, max_opacity = 0.7).add_to(base_map)
